Added new test classes

+++ src/main/java/pl/kurs/equationsolver/app/Main.java
@@ -6,6 +6,7 @@ import pl.kurs.equationsolver.services.*;
 import org.springframework.context.annotation.AnnotationConfigApplicationContext;
 import org.springframework.context.annotation.ComponentScan;
 
+import java.math.BigDecimal;
 import java.util.Scanner;
 
 @ComponentScan(basePackages = "pl.kurs")
@@ -19,7 +20,7 @@ public class Main {
 
         try {
 
-            int evaluateExpression = solvingService.evaluateExpression(args[0]);
+            BigDecimal evaluateExpression = solvingService.evaluateExpression(args[0]);
             System.out.println("wynik:" + evaluateExpression);
         } catch (InvalidEquationFormatException | UnknownOperatorException e) {
             e.printStackTrace();
