+++ src/main/java/pl/kurs/equationsolver/services/ISolvingService.java
@@ -3,7 +3,9 @@ package pl.kurs.equationsolver.services;
 import pl.kurs.equationsolver.exceptions.InvalidEquationFormatException;
 import pl.kurs.equationsolver.exceptions.UnknownOperatorException;
 
+import java.math.BigDecimal;
+
 public interface ISolvingService {
 
-    Integer evaluateExpression(String expression) throws InvalidEquationFormatException, UnknownOperatorException;
+    BigDecimal evaluateExpression(String expression) throws InvalidEquationFormatException, UnknownOperatorException;
 }
