+++ src/test/java/pl/kurs/equationsolver/services/SubtractTest.java
@@ -3,6 +3,8 @@ package pl.kurs.equationsolver.services;
 import org.junit.Assert;
 import org.junit.Test;
 
+import java.math.BigDecimal;
+
 import static org.junit.Assert.*;
 
 public class SubtractTest {
@@ -12,12 +14,12 @@ public class SubtractTest {
     @Test
     public void getResult() {
 
-        int a = 50;
-        int b = 10;
+        BigDecimal a = BigDecimal.valueOf(50);
+        BigDecimal b = BigDecimal.valueOf(10);
 
-        int expectedResult = 40;
+        BigDecimal expectedResult = BigDecimal.valueOf(40);
 
-        int result = subtract.getResult(a, b);
+        BigDecimal result = subtract.getResult(a, b);
 
 
         Assert.assertEquals(expectedResult, result);
